+++ bindings/nostr-sdk-ffi/src/protocol/types/image.rs
@@ -2,44 +2,30 @@
 // Copyright (c) 2023-2024 Rust Nostr Developers
 // Distributed under the MIT software license
 
-use std::ops::Deref;
+use uniffi::Record;
 
-use uniffi::Object;
-
-#[derive(Debug, PartialEq, Eq, Hash, Object)]
-#[uniffi::export(Debug, Eq, Hash)]
+#[derive(Record)]
 pub struct ImageDimensions {
-    inner: nostr::ImageDimensions,
-}
-
-impl Deref for ImageDimensions {
-    type Target = nostr::ImageDimensions;
-
-    fn deref(&self) -> &Self::Target {
-        &self.inner
-    }
+    /// Width
+    pub width: u64,
+    /// Height
+    pub height: u64,
 }
 
 impl From<nostr::ImageDimensions> for ImageDimensions {
     fn from(inner: nostr::ImageDimensions) -> Self {
-        Self { inner }
+        Self {
+            width: inner.width,
+            height: inner.height,
+        }
     }
 }
 
-#[uniffi::export]
-impl ImageDimensions {
-    #[uniffi::constructor]
-    pub fn new(width: u64, height: u64) -> Self {
+impl From<ImageDimensions> for nostr::ImageDimensions {
+    fn from(inner: ImageDimensions) -> Self {
         Self {
-            inner: nostr::ImageDimensions { width, height },
+            width: inner.width,
+            height: inner.height,
         }
-    }
-
-    pub fn width(&self) -> u64 {
-        self.inner.width
-    }
-
-    pub fn height(&self) -> u64 {
-        self.inner.height
     }
 }